server/requirements.txt
Django==4.2.3
gunicorn==21.2.0
psycopg2-binary==2.9.6
beautifulsoup4==4.12.2
openai==0.27.8
requests==2.31.0
tiktoken==0.4.0
nltk==3.8.1
python-dotenv==1.0.0
djangorestframework==3.14.0
django-allauth==0.58.2
djangorestframework-simplejwt==5.3.0
